Choropleth Map
Interactive Visual Analytics H590
For this project we have used 2015 American Community Survey (ACS) as
our primary dataset. The ACS dataset is annually compiled by the US
Census, and designed to provide broad demographic and socio-economic
characteristics of the US population.

This project aims to visualize 13 variables of population and census
data using a combination of choropleths, bar charts, pie charts,
scatterplots and multiple maps. We have provided users with the ability
to view the variable distribution of all states or the ability to
select a state and obtain detailed information of different variables
for each state. Our analysis has provided the following insights. The
native population is highest in southern states such as New Mexico,
Texas, Oklahoma & Arizona. A quick google search revealed that the
region of Oklahoma had settled occupation that dates back to a previous
time in history. This can be attributed to high number of natives in
this region. Hopefully by using this visualization you can help us
gather further insights into the data.